dom/filehandle/test/test_getFile.html

<!--
  Any copyright is dedicated to the Public Domain.
  http://creativecommons.org/publicdomain/zero/1.0/
-->
<html>
<head>
  <title>File Handle Test</title>

  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />

  <script type="text/javascript;version=1.7">
  function testSteps()
  {
    for each (let fileStorage in fileStorages) {
      let request = getFileHandle(fileStorage.key, "test.txt");
      request.onerror = errorHandler;
      request.onsuccess = grabEventAndContinueHandler;
      let event = yield undefined;

      let fileHandle = event.target.result;
      fileHandle.onerror = errorHandler;

      request = fileHandle.getFile();
      ok(request instanceof DOMRequest, "Correct interface");
      ok(!(request instanceof FileRequest), "Correct interface");
      ok(!('lockedFile' in request), "Property should not exist");
      ok(request.lockedFile === undefined, "Property should not exist");
      ok(!('onprogress' in request), "Property should not exist");
      ok(request.onprogress === undefined, "Property should not exist");
    }

    finishTest();
    yield undefined;
  }
  </script>
  <script type="text/javascript;version=1.7" src="helpers.js"></script>

</head>

<body onload="runTest();"></body>

</html>
